day3/abstraction-level-7/requirements.txt
fastapi
uvicorn
typer
pyyaml